JDK 1.5 でコンパイルした際のエラーを修正しました。
Update  issue 31

git-svn-id: http://aipo.googlecode.com/svn/trunk@211 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ portlets/blog/src/main/java/com/aimluck/eip/blog/BlogWordSelectData.java
@@ -138,13 +138,13 @@ public class BlogWordSelectData extends ALAbstractSelectData<DataRow, DataRow> {
       DataRow dataRow = obj;
 
       Integer entry_id = dataRow.getValue(EipTBlogEntry.ENTRY_ID_PK_COLUMN);
-      long ower_id = dataRow.getValue(EipTBlogEntry.OWNER_ID_COLUMN);
+      Integer ower_id = dataRow.getValue(EipTBlogEntry.OWNER_ID_COLUMN);
 
       BlogEntryResultData rd = new BlogEntryResultData();
       rd.initField();
       rd.setEntryId(entry_id.longValue());
       rd.setOwnerId(ower_id);
-      rd.setOwnerName(BlogUtils.getUserFullName((int) ower_id));
+      rd.setOwnerName(BlogUtils.getUserFullName(ower_id));
 
       rd.setTitle(ALCommonUtils.compressString((String) dataRow
         .get(EipTBlogEntry.TITLE_COLUMN), getStrLength()));
